services/plugin-system/app/core/database.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()

def get_mongo_client() -> MongoClient:
    """获取MongoDB客户端连接"""
    mongo_url = os.getenv("MONGO_URL", "mongodb://localhost:27017")
    client = MongoClient(mongo_url)
    return client