services/plugin-system/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
pyyaml
jsonschema
